+++ validator.go
@@ -36,7 +36,7 @@ type (
 		// Validates passed data and returns a map of possible validation errors happened on every field with failed validation.
 		//
 		// If no errors found, output will be nil
-		Validate(input interface{}) interface{}
+		Validate(input interface{}, translator ...func(string) string) interface{}
 		// Decrypts errors returned from gin's Bind process and returns proper error messages
 		//
 		// If returnUnmarshalErrorContext is true (default is true), if an error happened when
@@ -65,29 +65,27 @@ func (err sliceValidationError) Error() string {
 	return "error"
 }
 
+func getFormattedErrorMessage(message string, fieldName string, value interface{}, options option) string {
+	for key, value := range options {
+		message = strings.ReplaceAll(message, "$"+key, value)
+	}
+	return strings.ReplaceAll(strings.ReplaceAll(message, "$field", fieldName), "$value", fmt.Sprint(value))
+}
+
 // Formats and returns error message associated with passed ruleKey
-func getErrorMessage(fieldName string, ruleKey string, value interface{}, options option, messages Messages, specificMessage Messages, defaultErrorMessages map[string]string) string {
+func getRawErrorMessage(ruleKey string, messages Messages, specificMessage Messages, defaultErrorMessages map[string]string) string {
 	// Search for error message in specific messages for the rule
 	if out, ok := specificMessage[gStrings.SnakeCase(ruleKey)]; len(specificMessage) != 0 && ok {
-		for key, value := range options {
-			out = strings.ReplaceAll(out, "$"+key, value)
-		}
-		return strings.ReplaceAll(strings.ReplaceAll(out, "$field", fieldName), "$value", fmt.Sprint(value))
+		return out
 	}
 
 	// Search for error message in general messages for the validator
 	if out, ok := messages[ruleKey]; ok {
-		for key, value := range options {
-			out = strings.ReplaceAll(out, "$"+key, value)
-		}
-		return strings.ReplaceAll(strings.ReplaceAll(out, "$field", fieldName), "$value", fmt.Sprint(value))
+		return out
 	} else {
 		// If no error message found, search if there is a default error message for rule error
 		if defaultErrorMessage, ok := defaultErrorMessages[ruleKey]; ok {
-			for key, value := range options {
-				defaultErrorMessage = strings.ReplaceAll(defaultErrorMessage, "$"+key, value)
-			}
-			return strings.ReplaceAll(strings.ReplaceAll(defaultErrorMessage, "$field", fieldName), "$value", fmt.Sprint(value))
+			return defaultErrorMessage
 		} else {
 			// If no error message found, return that error message doesn't exist
 			return fmt.Sprintf("error happened but no error message exists on '%s' rule key", ruleKey)
@@ -95,9 +93,13 @@ func getErrorMessage(fieldName string, ruleKey string, value interface{}, option
 	}
 }
 
-func (o *validatorS) Validate(input interface{}) interface{} {
+func (o *validatorS) Validate(input interface{}, translator ...func(string) string) interface{} {
 	for reflect.ValueOf(input).Kind() == reflect.Ptr {
 		input = reflect.ValueOf(input).Elem().Interface()
+	}
+	var t func(string) string = nil
+	if len(translator) != 0 {
+		t = translator[0]
 	}
 
 	output := map[string]interface{}{}
@@ -122,7 +124,12 @@ func (o *validatorS) Validate(input interface{}) interface{} {
 				if o.messages != nil {
 					m = *o.messages
 				}
-				halfOutput = append(halfOutput, getErrorMessage(fieldName, failKey, onKeyInput, ruleSet.getOption(failKey), m, sm, defaultValidatorErrorMessages))
+				message := getRawErrorMessage(failKey, m, sm, defaultValidatorErrorMessages)
+				if t != nil {
+					message = t(message)
+				}
+				message = getFormattedErrorMessage(message, fieldName, onKeyInput, ruleSet.getOption(failKey))
+				halfOutput = append(halfOutput, message)
 			}
 		}
 
@@ -169,7 +176,7 @@ func (o *validatorS) Validate(input interface{}) interface{} {
 				}
 
 				if ruleSet.hasDeepValidator() && output[fieldName] == nil && (mapRule(value) || structRule(value) || sliceRule(value)) {
-					data := ruleSet.validateDeepValidator(value)
+					data := ruleSet.validateDeepValidator(value, t)
 
 					if reflect.ValueOf(data).IsValid() && reflect.ValueOf(data).Len() != 0 {
 						output[fieldName] = data
@@ -180,7 +187,7 @@ func (o *validatorS) Validate(input interface{}) interface{} {
 					valueOnKeyInput = reflect.ValueOf(valueOnKeyInput.Interface())
 					for i := 0; i < valueOnKeyInput.Len(); i++ {
 						element := valueOnKeyInput.Index(i)
-						errors := ruleSet.validateChildrenValidator(element.Interface())
+						errors := ruleSet.validateChildrenValidator(element.Interface(), t)
 						if reflect.ValueOf(errors).IsValid() && reflect.ValueOf(errors).Len() != 0 {
 							if _, ok := output[fieldName]; !ok {
 								output[fieldName] = map[string]interface{}{}
@@ -228,7 +235,7 @@ func (o *validatorS) Validate(input interface{}) interface{} {
 				}
 
 				if ruleSet.hasDeepValidator() && output[fieldName] == nil && (mapRule(value) || structRule(value) || sliceRule(value)) {
-					data := ruleSet.validateDeepValidator(value)
+					data := ruleSet.validateDeepValidator(value, t)
 
 					if reflect.ValueOf(data).IsValid() && reflect.ValueOf(data).Len() != 0 {
 						output[fieldName] = data
@@ -239,7 +246,7 @@ func (o *validatorS) Validate(input interface{}) interface{} {
 					valueOnKeyInput = reflect.ValueOf(valueOnKeyInput.Interface())
 					for i := 0; i < valueOnKeyInput.Len(); i++ {
 						element := valueOnKeyInput.Index(i)
-						errors := ruleSet.validateChildrenValidator(element.Interface())
+						errors := ruleSet.validateChildrenValidator(element.Interface(), t)
 						if reflect.ValueOf(errors).IsValid() && reflect.ValueOf(errors).Len() != 0 {
 							if _, ok := output[fieldName]; !ok {
 								output[fieldName] = map[string]interface{}{}
@@ -267,7 +274,7 @@ func (o *validatorS) Validate(input interface{}) interface{} {
 			if o.rule.hasChildrenValidator() {
 				for i := 0; i < inputValue.Len(); i++ {
 					element := inputValue.Index(i)
-					errors := o.rule.validateChildrenValidator(element.Interface())
+					errors := o.rule.validateChildrenValidator(element.Interface(), t)
 
 					if reflect.ValueOf(errors).IsValid() && reflect.ValueOf(errors).Len() != 0 {
 						if _, ok := output[strconv.Itoa(i)]; !ok {
@@ -278,7 +285,7 @@ func (o *validatorS) Validate(input interface{}) interface{} {
 			}
 		default:
 			if o.rule.hasDeepValidator() {
-				errors := o.rule.validateDeepValidator(input)
+				errors := o.rule.validateDeepValidator(input, t)
 				if reflect.ValueOf(errors).IsValid() && reflect.ValueOf(errors).Len() != 0 {
 					return errors
 				}
@@ -334,7 +341,10 @@ func decryptPath(path string, v Validator, errorField playgroundValidator.FieldE
 				if r.getName() != "" {
 					fieldName = r.getName()
 				}
-				return getErrorMessage(fieldName, errorField.Tag(), errorField.Value(), r.getOption(errorField.Tag()), *v.getMessages(), r.getSpecificMessages(), defaultValidatorErrorMessages)
+				message := getRawErrorMessage(errorField.Tag(), *v.getMessages(), r.getSpecificMessages(), defaultValidatorErrorMessages)
+				// Do not add translator, translator is for Validate process
+				message = getFormattedErrorMessage(message, fieldName, errorField.Value(), r.getOption(errorField.Tag()))
+				return message
 			} else {
 				panic("error structure does not match with validator structure")
 			}